+++ packages/sitecore-jss/src/layout/graphql-layout-service.ts
@@ -1,9 +1,13 @@
 import { LayoutServiceBase } from './layout-service';
 import { LayoutServiceData } from './models';
-import { GraphQLClient, GraphQLRequestClient } from '../graphql-request-client';
+import {
+  GraphQLClient,
+  GraphQLRequestClient,
+  GraphQLRequestClientConfig,
+} from '../graphql-request-client';
 import debug from '../debug';
 
-export type GraphQLLayoutServiceConfig = {
+export type GraphQLLayoutServiceConfig = Pick<GraphQLRequestClientConfig, 'retries'> & {
   /**
    * Your Graphql endpoint
    */
@@ -79,6 +83,7 @@ export class GraphQLLayoutService extends LayoutServiceBase {
     return new GraphQLRequestClient(this.serviceConfig.endpoint, {
       apiKey: this.serviceConfig.apiKey,
       debugger: debug.layout,
+      retries: this.serviceConfig.retries,
     });
   }